requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
filelock==3.9.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
pandas==2.0.2
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.1
six==1.16.0
sympy==1.11.1
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
tqdm==4.65.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
